Closes #15
Attacks occured whether or not the enemy was dead, as such so did
counterattacks. This combined with the high chance of daggers
counterattacking meant that an enemy and player continually
counterattacking could end in a nearly endless loop and stack overflow.
Fix simply stops attacks on things that are already dead.
Closes issue #15